restructure architecture and interfaces
The old design was not flexible regarding system creation and mixing
components. A new design approach is chosen where key components are
defined by base classes or abstract base classes and the final system is
created by composition. This approach is flexibil and can be extended
easily from here. The gcode module is removed for now but maybe re
initiated after the optimization module works or out sourced.

Chages:
- decouple optical system from stage
- create distinc visualization object to separate tasks
- create generic source as an adapter between system and differnt
sources
- implemented different sources
- restructured surface module with base class and inheritance
- new optical system is a composition allowing more flexibility
- create new constants for zero-vectors and more